requirements.txt
streamlit>=1.44.0
pypdf>=6.10.0
openai>=2.0.0
anthropic>=0.67.0
google-genai>=1.36.0
